feat: utilize `MediaSession` API: PiP, etc
On Android this should let us utilize Picture-in-Picture
when minimizing the app to the background.
This also enables mic and camera toggles
on the said Picture-in-Picture window.

On Windows on Delta Chat Desktop this does not appear
to have much of an effect currently.
This will not `requestPictureInPicture()` automatically
(but you can try it manually to see what it looks like).
Additionally, the `MediaSession` seems to be limited on Electron
in general:
https://github.com/electron/electron/issues/25539.

But this has a potential for exploration if we add an explicit
"Picture-in-Picture" button.
Such a native Chromium Picture-in-Picture window
looks basically equivalent to what we have implemented manually
in this app.

Anyways, how well this works depends on the platform
that the app is running in, so let's do our part.